Rename test_replace_all_html
