+++ crates/wysiwyg/src/tests.rs
@@ -21,7 +21,7 @@ pub mod test_links;
 pub mod test_lists;
 pub mod test_menu_state;
 pub mod test_paragraphs;
-pub mod test_replace_all_html;
+pub mod test_replace_all_content;
 pub mod test_selection;
 pub mod test_to_markdown;
 pub mod test_to_raw_text;
